feat(frontend): single DuckDB tile with MotherDuck/DuckLake switcher
