+++ packages/frontend/src/components/ProjectConnection/DbtSettingsForm.tsx
@@ -25,7 +25,6 @@ import { BigQuerySchemaInput } from './WarehouseForms/BigQueryForm';
 import { ClickhouseSchemaInput } from './WarehouseForms/ClickhouseForm';
 import { DatabricksSchemaInput } from './WarehouseForms/DatabricksForm';
 import { DuckdbSchemaInput } from './WarehouseForms/DuckdbForm';
-import { DucklakeSchemaInput } from './WarehouseForms/DucklakeForm';
 import { PostgresSchemaInput } from './WarehouseForms/PostgresForm';
 import { RedshiftSchemaInput } from './WarehouseForms/RedshiftForm';
 import { SnowflakeSchemaInput } from './WarehouseForms/SnowflakeForm';
@@ -152,8 +151,6 @@ const DbtSettingsForm: FC<DbtSettingsFormProps> = ({
                 return AthenaSchemaInput;
             case WarehouseTypes.DUCKDB:
                 return DuckdbSchemaInput;
-            case WarehouseTypes.DUCKLAKE:
-                return DucklakeSchemaInput;
             default: {
                 return assertUnreachable(
                     warehouseType,
